Modify interactive hispeed_freq
Sync with QCOM: The interactive governor is configured with a hispeed
frequency of 1.5 GHz. This causes L2 to be bumped up to 1.5 GHz as well,
which requires a turbo CX voltage and affects power. To save power when
the hispeed frequency is triggered, change the hispeed frequency to
1.2 GHz, which has the highest L2 frequency that can be supported by a
nominal CX voltage.